# src-cloud/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "my-todo",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/todos": {
      "get": {
        "tags": [
          "domains::todos::controller"
        ],
        "operationId": "find_all",
        "responses": {
          "200": {
            "description": "Todos found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Todos not found"
          }
        }
      },
      "post": {
        "tags": [
          "domains::todos::controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTodo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created Todo successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "Todo couldn't be created"
          }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "tags": [
          "domains::todos::controller"
        ],
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "todo id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "Todo not found"
          }
        }
      },
      "delete": {
        "tags": [
          "domains::todos::controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "todo id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "todo successfully deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "204": {
            "description": "todo not deleted"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "domains::todos::controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "todo id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTodo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "todo successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            }
          },
          "404": {
            "description": "todo not updated"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateTodo": {
        "type": "object",
        "required": [
          "text"
        ],
        "properties": {
          "text": {
            "type": "string"
          }
        }
      },
      "Todo": {
        "type": "object",
        "required": [
          "id",
          "text",
          "completed"
        ],
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": "string"
          }
        }
      },
      "UpdateTodo": {
        "type": "object",
        "properties": {
          "completed": {
            "type": "boolean",
            "nullable": true
          },
          "text": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}